requirements.txt
gitdb2==2.0.5
GitPython==2.1.11
ipykernel==5.1.0
ipyparallel==6.2.3
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
lxml==4.5.1
opencv-python==3.4.5.20
Pillow==5.4.1
tqdm==4.31.1
git+git://github.com/matterport/Mask_RCNN.git@master
git+git://github.com/lorenzo-park/DocParser@master
BentoML==0.12.1
tensorflow-gpu==2.4.1
scikit-image==0.14.2